Add MultiSet{T}(), built on Accumulator{T,Int}().
A `MultiSet{T}` is very much like an `Accumulator{T,Int}`, but it
exposes a single-value-based interface. The fact that it is storing
counts under the hood is completely hidden from the interface.

For example, iterating the MultiSet returns each item in the MultiSet,
repeated per its count, as if the multiset _actually_ contained
duplicate items:
```julia
julia> show(collect(MultiSet([1,1,1,2,2])))
[2, 2, 1, 1, 1]
```
whereas the accumulator presents its count-based representation to the
user:
```julia
julia> show(collect(MultiSet([1,1,1,2,2]).data))
[2 => 2, 1 => 3]
```

For more info on MultiSets, see:
https://en.wikipedia.org/wiki/Multiset