implement new inner constructor
implement new inner constructor which takes in the starting index (typically 1), the length of the currently filled portion of the buffer, and the preconstructed buffer. Users will typically use the other constructors. Direct usage of the inner constructor should be considered primarily an internal interface.
Also fixes the outer constructors to match the new inner constructor.