Refactor popfirst!(::PriorityQueue)
`PriorityQueue` has been introduced as "acting like a `Dict`".
Therefore, it may be considered as a collection of `key => priority`
pairs, as indicated by `first(pq)` or iteration over it. However,
`popfirst!(pq)` has returned only the key but not the pair, such that
the return type of `popfirst!` did not match `eltype(pq)`. This commit
fixes that by replacing `popfirst!` with `dequeue_pair!`.

This is a breaking change.